% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/abbreviation_internal.R
\name{abbreviation_internal}
\alias{abbreviation_internal}
\title{Internal abbreviation marker, marks abbreviations with an underscore behind.
Useful if \code{parsing_option} 1 is needed, but some abbreviations need \code{parsing_option} 2.}
\usage{
abbreviation_internal(string, abbreviations = NULL)
}
\arguments{
\item{string}{A string (for example names of a data frame).}

\item{abbreviations}{character with (uppercase) abbreviations. This marks
abbreviations with an underscore behind (in front of the parsing).
Useful if \code{parsing_option} 1 is needed, but some abbreviations need \code{parsing_option} 2.}
}
\value{
A character vector.
}
\description{
Internal abbreviation marker, marks abbreviations with an underscore behind.
Useful if \code{parsing_option} 1 is needed, but some abbreviations need \code{parsing_option} 2.
}
\author{
Malte Grosser, \email{malte.grosser@gmail.com}
}
\keyword{utilities}
