% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/caseconverters.R
\name{caseconverter_shortcuts}
\alias{caseconverter_shortcuts}
\alias{to_snake_case}
\alias{to_small_camel_case}
\alias{to_big_camel_case}
\alias{to_screaming_snake_case}
\alias{to_parsed_case}
\title{caseconverter}
\usage{
to_snake_case(string)

to_small_camel_case(string)

to_big_camel_case(string)

to_screaming_snake_case(string)

to_parsed_case(string)
}
\arguments{
\item{string}{A string.}
}
\value{
A character vector according the specified case.
}
\description{
These functions allow you to parse and convert a string to a specific case.
}
\examples{
strings <- c("this Is a Strange_string", "AND THIS ANOTHER_One", NA)
to_snake_case(strings)
to_small_camel_case(strings)
to_big_camel_case(strings)
to_screaming_snake_case(strings)

}
