% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parsing_helpers.R
\name{parse1_pat_cap_smalls}
\alias{parse1_pat_cap_smalls}
\alias{parse2_pat_caps2}
\alias{parse3_pat_cap_lonely}
\alias{parse4_separate_non_characters}
\alias{parse5_mark_digits}
\title{Parsing helpers for parsed_case_internal}
\usage{
parse1_pat_cap_smalls(string)

parse2_pat_caps2(string)

parse3_pat_cap_lonely(string)

parse4_separate_non_characters(string)

parse5_mark_digits(string)
}
\arguments{
\item{string}{A string.}
}
\value{
A character vector separated by underscores, containing the parsed string.
}
\description{
Parsing helpers for parsed_case_internal
}
\author{
Malte Grosser, \email{malte.grosser@gmail.com}
}
\keyword{utilities}
