% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocess_internal.R
\name{preprocess_internal}
\alias{preprocess_internal}
\title{Internal function that replaces regex matches with underscores}
\usage{
preprocess_internal(string, sep_in)
}
\arguments{
\item{string}{A string.}

\item{sep_in}{(short for separator input) A regex supplied as a character (if not \code{NULL}), which will be wrapped internally
into \code{stringr::regex()}. All matches will be replaced by underscores (additionally to 
\code{"_"} and \code{" "}, for which this is always true). Underscores can later turned into another separator via \code{postprocess}.}
}
\value{
A character containing the parsed string.
}
\description{
Internal function that replaces regex matches with underscores
}
\author{
Malte Grosser, \email{malte.grosser@gmail.com}
}
\keyword{utilities}
