% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/protect_internal.R
\name{protect_internal}
\alias{protect_internal}
\title{Internal protect helper (deletes the separators around protected regex matches).}
\usage{
protect_internal(string, protect = NULL)
}
\arguments{
\item{string}{A string (for example names of a data frame).}

\item{protect}{A string which is a valid \code{stringr::regex()}. Matches within the output
won't have any "_" (or artifacts of \code{preprocess}) beside. Note that \code{preprocess} has a higher precedence than protect, 
which means that it doesn't make sense to protect sth. which is already replaced
via \code{preprocess}.}
}
\value{
A character vector.
}
\description{
Internal protect helper (deletes the separators around protected regex matches).
}
\author{
Malte Grosser, \email{malte.grosser@gmail.com}
}
\keyword{utilities}
