% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/relevant.R
\name{relevant}
\alias{relevant}
\title{Internal helper for "lower_upper", "upper_lower".
This helper returns a logical vector with TRUE for the 
first and every second string of those which contain 
an alphabetic character}
\usage{
relevant(string)
}
\arguments{
\item{string}{A string (for example names of a data frame).}
}
\value{
A logical vector.
}
\description{
Internal helper for "lower_upper", "upper_lower".
This helper returns a logical vector with TRUE for the 
first and every second string of those which contain 
an alphabetic character
}
\author{
Malte Grosser, \email{malte.grosser@gmail.com}
}
\keyword{utilities}
