% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/replace_special_characters_internal.R
\name{replace_special_characters_internal}
\alias{replace_special_characters_internal}
\title{Internal helper to replace special characters.}
\usage{
replace_special_characters_internal(string)
}
\arguments{
\item{string}{A string (for example names of a data frame).}
}
\value{
A character vector.
}
\description{
Internal helper to replace special characters.
}
\author{
Malte Grosser, \email{malte.grosser@gmail.com}
}
\keyword{utilities}
