% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/to_any_case.R
\name{to_any_case_dev}
\alias{to_any_case_dev}
\title{Functions to convert column names to snake_case}
\usage{
to_any_case_dev(string, case = c("snake", "small_camel", "big_camel"),
  preprocess = "\\\\s+", postprocess = "_")
}
\arguments{
\item{string}{Character string indicating column names of a data.frame.}

\item{case}{Character string ("snake", "small_camel" or "big_camel"), indicating 
case should be the target case, that the string should be converted into.}

\item{preprocess}{Character string that will be wrapped internally into stringr::regex. All matches will be
treated like underscore. This is useful and gives flexibility to customize conversion
of strings that contain also dots, hyphens and/or other special characters.}

\item{postprocess}{Character string indicating column names of a data.frame.}
}
\value{
character
}
\description{
Functions to convert column names to snake_case
}
\examples{
camelCases <- c("smallCamelCase", "BigCamelCase", "mixed_Case", "snake_case", "_camel_case__")
to_ana_case(camelCases)
to_any_case_dev("fsdf.d-sf", preprocess = "\\\\.|-")

}
\author{
Malte Grosser, \email{malte.grosser@gmail.com}
}
\keyword{utilities}
