% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/to_parsed_case.R
\name{to_parsed_case}
\alias{to_parsed_case}
\title{Functions to convert column names to snake_case}
\usage{
to_parsed_case(string)
}
\arguments{
\item{string}{Character string indicating column names of a data.frame.}
}
\value{
Vector of character strings, parsed for conversion to snake_case, separated by "_"
}
\description{
Functions to convert column names to snake_case
}
\examples{
camelCases <- c("smallCamelCase", "BigCamelCase", "mixed_Case", "snake_case", "_camel_case__")
to_parsed_case(camelCases)
to_parsed_case(names(iris))
to_parsed_case(c(".", "_"))
to_parsed_case(NA)
to_parsed_case(NULL)
to_parsed_case(character(0))
to_parsed_case("   N A")

}
\author{
Malte Grosser, \email{malte.grosser@gmail.com}
}
\keyword{utilities}
