% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/to_small_camel_case.R
\name{to_small_camel_case}
\alias{to_small_camel_case}
\title{Functions to convert column names to smallCamelCase}
\usage{
to_small_camel_case(snake_case)
}
\arguments{
\item{camelcases}{Character string indicating column names of a data.frame.}
}
\value{
Vector of character strings in smallCamelCase
}
\description{
Functions to convert column names to smallCamelCase
}
\examples{
camelCases <- c("smallCamelCase", "BigCamelCase", "mixed_Case", "snake_case", "_camel_case__")
to_small_camel_case(camelCases)
to_small_camel_case(names(iris))
to_small_camel_case(c(".", "_"))
to_small_camel_case(NA)
to_small_camel_case(NULL)
to_small_camel_case(character(0))
to_small_camel_case("   N A")

}
\author{
Malte Grosser, \email{malte.grosser@gmail.com}
}
\keyword{utilities}
