% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/to_snake_case.R
\name{to_snake_case_dev}
\alias{to_snake_case_dev}
\title{Functions to convert column names to snake_case}
\usage{
to_snake_case_dev(camelcases)
}
\arguments{
\item{camelcases}{Character string indicating column names of a data.frame.}
}
\value{
Vector of character strings in snake_case
}
\description{
Functions to convert column names to snake_case
}
\examples{
camelCases <- c("smallCamelCase", "BigCamelCase", "mixed_Case", "snake_case", "_camel_case__")
to_snake_case(camelCases)
to_snake_case(names(iris))
to_snake_case(c(".", "_"))
to_snake_case(NA)
to_snake_case(NULL)
to_snake_case(character(0))
to_snake_case("   N A")

}
\author{
Malte Grosser, \email{malte.grosser@gmail.com}
}
\keyword{utilities}
