% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/to_snake_case_internal.R
\name{to_snake_case_internal}
\alias{to_snake_case_internal}
\title{Functions to convert column names to snake_case}
\usage{
to_snake_case_internal(string, preprocess = NULL)
}
\arguments{
\item{string}{Character string indicating column names of a data.frame.}

\item{preprocess}{Character string that will be wrapped internally into stringr::regex. All matches will be
treated like underscore. This is useful and gives flexibility to customize conversion
of strings that contain also dots, hyphens and/or other special characters.}
}
\value{
Vector of character strings in snake_case
}
\description{
Functions to convert column names to snake_case
}
\author{
Malte Grosser, \email{malte.grosser@gmail.com}
}
\keyword{utilities}
