% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocess_internal.R
\name{preprocess_internal}
\alias{preprocess_internal}
\title{Internal function that replaces regex matches with underscores}
\usage{
preprocess_internal(string, preprocess)
}
\arguments{
\item{string}{A string.}

\item{preprocess}{Character string that will be wrapped internally into stringr::regex. 
All matches will be padded with underscores.}
}
\value{
A character containing the parsed string.
}
\description{
Internal function that replaces regex matches with underscores
}
\author{
Malte Grosser, \email{malte.grosser@gmail.com}
}
\keyword{utilities}
