% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/replace_special_characters_internal.R
\name{replace_special_characters_internal}
\alias{replace_special_characters_internal}
\title{Internal helper to replace special characters.}
\usage{
replace_special_characters_internal(string, transliterations, case)
}
\arguments{
\item{string}{A string (for example names of a data frame).}

\item{transliterations}{A character vector (if not \code{NULL}). The entries of this argument
need to be elements of \code{stringi::stri_trans_list()} (like "Latin-ASCII", which is often useful) or names of lookup tables (currently
only "german" is supported). In the order of the entries the letters of the input
 string will be transliterated via \code{stringi::stri_trans_general()} or replaced via the 
 matches of the lookup table. When named character elements are supplied as part of `transliterations`, anyting that mateches the names is replaced by the corresponding value.
You should use this feature with care in case of \code{case = "parsed"}, \code{case = "internal_parsing"} and 
\code{case = "none"}, since for upper case letters, which have transliterations/replacements
 of length 2, the second letter will be transliterated to lowercase, for example Oe, Ae, Ss, which
 might not always be what is intended. In this case you can make usage of the option to supply named elements and specify the transliterations yourself.}

\item{case}{Length one character, from the input options of \code{to_any_case}.}
}
\value{
A character vector.
}
\description{
Internal helper to replace special characters.
}
\author{
Malte Grosser, \email{malte.grosser@gmail.com}
}
\keyword{utilities}
