% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/replace_special_characters_internal.R
\name{replace_special_characters_internal}
\alias{replace_special_characters_internal}
\title{Internal helper to replace special characters.}
\usage{
replace_special_characters_internal(string, replace_special_characters, case)
}
\arguments{
\item{string}{A string (for example names of a data frame).}

\item{replace_special_characters}{Logical, if \code{TRUE}, special characters 
will be translated to characters which are more likely to be understood by 
different programs. For example german umlauts will be translated to ae, oe, ue etc.}

\item{case}{Length one character, from the input options of \code{to_any_case}.}
}
\value{
A character vector.
}
\description{
Internal helper to replace special characters.
}
\author{
Malte Grosser, \email{malte.grosser@gmail.com}
}
\keyword{utilities}
