% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/to_any_case.R
\name{to_any_case}
\alias{to_any_case}
\title{Functions to convert column names to snake_case}
\usage{
to_any_case(string, case = c("snake", "small_camel", "big_camel",
  "screaming_snake"), preprocess = NULL, postprocess = NULL, prefix = "",
  postfix = "", replace_special_characters = FALSE)
}
\arguments{
\item{string}{Character string indicating column names of a data.frame.}

\item{case}{Character string ("snake", "small_camel", "big_camel" or "screaming_snake"), indicating 
case should be the target case, that the string should be converted into.}

\item{preprocess}{Character string that will be wrapped internally into stringr::regex. All matches will be
treated like underscore. This is useful and gives flexibility to customize conversion
of strings that contain also dots, hyphens and/or other special characters.}

\item{postprocess}{Character string that will be the separator between different words. When `case == "snake"`
it will replace the underscore.}

\item{prefix}{prefix}

\item{postfix}{postfix}

\item{replace_special_characters}{if `TRUE`, special characters will be translated to characters which are more likely to be understood by different programs. For example german umlauts will be translated to ae, oe, ue etc.}
}
\value{
character
}
\description{
Functions to convert column names to snake_case
}
\examples{
camelCases <- c("smallCamelCase", "BigCamelCase", "mixed_Case", "snake_case", "_camel_case__")
to_any_case(camelCases, case = "snake")
to_any_case("fsdf.d-sf", case = "snake", preprocess = "\\\\.|-")

}
\author{
Malte Grosser, \email{malte.grosser@gmail.com}
}
\keyword{utilities}
