% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/to_parsed_case_internal.R
\name{to_parsed_case_internal}
\alias{to_parsed_case_internal}
\title{Functions to convert column names to snake_case}
\usage{
to_parsed_case_internal(string, preprocess = NULL)
}
\arguments{
\item{string}{A string.}

\item{preprocess}{Character string that will be wrapped internally into stringr::regex. 
All matches will be padded with underscores.}
}
\value{
A character vector separated by underscores, containing the parsed string.
}
\description{
Functions to convert column names to snake_case
}
\author{
Malte Grosser, \email{malte.grosser@gmail.com}
}
\keyword{utilities}
